requirements.txt
streamlit
requests
geopy
pandas
folium
streamlit-folium
openai
openrouteservice
python-dotenv